refactor(ds): extract ClientInspector to features/clients (5d.3)
Second of the per-feature inspector splits. Moves the L1729-1830
client render block + the bottom-anchored archive confirmation Dialog
out of the 1924-line InspectorPanel into a focused 283-line
feature-local component.

ClientInspector
---------------
- src/features/clients/components/ClientInspector.tsx
- Props: practiceId, entityId (the userId for the client)
- Consumes useUserDetail directly — owns its data fetch + mutation
- Owns editor state: activePersonEditor, isSavingPersonField,
  isArchivingPerson, isArchiveConfirmOpen, personAddressDraft
- Owns helpers: readAddressFromDetail, formatAddressSummary,
  openPersonEditor, handlePersonFieldUpdate, handlePersonStatusChange
- Renders its own loading skeleton + the archive confirmation Dialog
- Display text token cleanups along the way: text-input-placeholder →
  text-dim, text-red-* removed in favor of text-neg

InspectorPanel dispatcher update
--------------------------------
- Imports + delegates: <ClientInspector practiceId={...} entityId={...} />
  when entityType === 'client' and both ids are present
- Removed 5 useState slots + 1 readAddressFromDetail + 1
  formatAddressSummary + 1 openPersonEditor + 1 handlePersonFieldUpdate
  + 1 handlePersonStatusChange (all moved to ClientInspector)
- Removed setActivePersonEditor from the reset useEffect (state no
  longer exists here)
- Removed the client loading skeleton block (ClientInspector renders
  its own)
- Removed the bottom-anchored archive Dialog (now in ClientInspector)
- Dropped unused imports: UserDetailRecord type (moved), Address type
  (moved), AddressExperienceForm (moved), Dialog (moved)
- Net change: +51 lines in ClientInspector, -232 lines in InspectorPanel

ConversationInspector (5d.5) will need its own copy of the contact-
display fields (name/email/phone/relationship status) for the
conversation block's L1307-1315 area. Pure read access via useUserDetail
— no need to share editor state since conversation inspector doesn't
edit the person record.
